Update iOS Firebase bindings to 12.7 (#618)

+++ src/Firestore/Platforms/iOS/FirebaseFirestoreImplementation.cs
@@ -106,7 +106,7 @@ public sealed class FirebaseFirestoreImplementation : DisposableBase, IFirebaseF
     /// <inheritdoc/>
     public void UseEmulator(string host, int port)
     {
-        _firestore.UseEmulatorWithHost(host, (uint) port);
+        _firestore.UseEmulatorWithHost(host, port);
         Settings = new FirestoreSettings(Settings.Host);
     }
